US72282 - Move AWS modules out of this repo
