+++ test/al_mock.js
@@ -1,0 +1,14 @@
+
+const AIMS_TEST_CREDS = {
+    access_key_id: 'test-access-key-id',
+    secret_key: 'test-secret-key'
+};
+
+const TEST_AL_API = 'api.test.product.dev.alertlogic.com';
+const INGEST_TEST_URL = '/data/secmsgs';
+
+module.exports = {
+    TEST_AL_API : TEST_AL_API,
+    AIMS_TEST_CREDS : AIMS_TEST_CREDS,
+    INGEST_TEST_URL : INGEST_TEST_URL
+};
