+++ test/al_test.js
@@ -1,0 +1,22 @@
+
+const assert = require('assert');
+const rewire = require('rewire');
+const sinon = require('sinon');
+const m_aimsc = require('../al_servicec').AimsC;
+const cweMock = require('./al_mock');
+var servicecRewire = rewire('../al_servicec');
+var m_servicec = require('../al_servicec');
+
+describe('Unit Tests', function() {
+
+    describe('getAlAuth()', function() {
+
+        before(function() {
+        });
+
+        afterEach(function() {            
+        });
+
+    });
+    
+});
